requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
coverage==7.4.0
cryptography==41.0.7
daphne==4.1.0
defusedxml==0.8.0rc2
Django==4.2
django-debug-toolbar==4.2.0
django-debug-toolbar-force==0.2
django-filter==23.5
django-nine==0.2.7
django-ordered-model==3.7.4
django-redis==5.4.0
djangochannelsrestframework==1.2.0
djangorestframework==3.14.0
drf-nested-routers==0.93.5
drf-yasg==1.21.7
flake8==7.0.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
isort==5.13.2
kombu==5.3.4
Markdown==3.5.1
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
pyOpenSSL==24.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==24.1.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.2
